feat: add persistence
